+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.3.0] - 2024-07-04
+
+_Stable release based on [0.3.0-rc.1]._
+
+## [0.3.0-rc.1] - 2024-07-04
+
+### Added
+
+- Add message and messages components.
+
 ## [0.2.0] - 2024-07-04
 
 _Stable release based on [0.2.0-rc.1]._
@@ -50,6 +60,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[0.3.0]: https://https://github.com/internetguru/blade-components/compare/v0.2.0...v0.3.0
+[0.3.0-rc.1]: https://github.com/internetguru/blade-components/releases/tag/v0.2.0
 [0.2.0]: https://https://github.com/internetguru/blade-components/compare/v0.1.4...v0.2.0
 [0.2.0-rc.1]: https://github.com/internetguru/blade-components/releases/tag/v0.1.4
 [0.1.4]: https://https://github.com/internetguru/blade-components/compare/v0.1.3...v0.1.4